move 'coverage' folder up one directory
